correction to jpgtopng

+++ week4/jpg2png.sh
@@ -2,7 +2,11 @@
 for f in * ; do
   if [[ "$f" =~ ".jpg" ]] ; then
     newF=`echo "$f" | sed -e "s/\.jpg/\.png/"`
-    convert "$f" "$newF"
-    rm "$f"
+	if [ -f "$newF" ] ; then
+		echo "$newF already exists"	
+	else
+		convert "$f" "$newF"
+    	rm "$f"
+	fi
   fi
 done
